Extract CSP report classification into CspReport
The job's own work is three lines; everything else was stateless domain
logic about what a report means, which is what CspPolicy already models.
Its specs no longer need a job harness, WebMock and a production Rails.env
to ask whether a report is noise.

Also scope the fetch-error filter to an actual navigation. Failed to fetch
and Load failed are what browsers report for any rejected fetch, so the
filter was suppressing the genuine failures the retry UI was just built
for. A pagehide flag distinguishes them.
